refactor(api): move dashboards to modules/ as controller, model, service
